refactor: standardize import statements and update API client structures
