+++ unifi/network_site.go
@@ -2,11 +2,16 @@ package unifi
 
 import (
 	"context"
+	"fmt"
 
-	network "github.com/ubiquiti-community/go-unifi/client/network"
+	"github.com/ubiquiti-community/go-unifi/client/network"
 )
 
 func (c *ApiClient) ListSites(ctx context.Context) ([]network.SiteOverview, error) {
+	if c.network == nil {
+		return nil, fmt.Errorf("Network API is unavailable")
+	}
+
 	return FetchAll(ctx, func(offset int32) (*network.SiteOverviewPage, error) {
 		resp, err := c.network.client.GetSiteOverviewPageWithResponse(ctx, &network.GetSiteOverviewPageParams{
 			Offset: Ptr(offset),
