+++ unifi/protect_light.go
@@ -1,0 +1,47 @@
+package unifi
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/ubiquiti-community/go-unifi/client/protect"
+)
+
+func (c *ApiClient) ListLights(ctx context.Context) ([]protect.Light, error) {
+	if c.protect == nil {
+		return nil, fmt.Errorf("Protect API is unavailable")
+	}
+
+	resp, err := c.protect.client.GetV1LightsWithResponse(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	return *resp.JSON200, nil
+}
+
+func (c *ApiClient) GetLight(ctx context.Context, id protect.LightId) (*protect.Light, error) {
+	if c.protect == nil {
+		return nil, fmt.Errorf("Protect API is unavailable")
+	}
+
+	resp, err := c.protect.client.GetV1LightsIdWithResponse(ctx, id)
+	if err != nil {
+		return nil, err
+	}
+
+	return resp.JSON200, nil
+}
+
+func (c *ApiClient) UpdateLight(ctx context.Context, id protect.LightId, params protect.PatchV1LightsIdJSONRequestBody) (*protect.Light, error) {
+	if c.protect == nil {
+		return nil, fmt.Errorf("Protect API is unavailable")
+	}
+
+	resp, err := c.protect.client.PatchV1LightsIdWithResponse(ctx, id, params)
+	if err != nil {
+		return nil, err
+	}
+
+	return resp.JSON200, nil
+}
